- implement Anthill::ants method - fix "set_field" call in place of "_set_field" in finish and fail. - fix Anthill mojolicious plugin typo - fix Spawn of command that contain spaces (in full path name)

+++ t/01_basic.t
@@ -3,22 +3,24 @@ use Test::More;
 
 use_ok('Anthill');
 
-my $db ='temp.db';
-unlink $db if -e $db;
+my $db = $ENV{TEST_DBI} || 'temp.db';
+my $dbq = $db =~ s/'/\\'/gr;	#escape quotes
+my $is_sqlite = $db =~ /\.db$/;
+unlink $db if $is_sqlite && -e $db;
+
 my $anthill = Anthill->new({ dbh => $db });
 is ref($anthill), 'Anthill', 'new Anthill';
 my $dbh = $anthill->dbh;
 is ref($dbh), 'DBI::db', 'anthill db handle';
 #deploying
 ok $dbh->do( Anthill->deploy_script( $dbh ) ), 'deploy anthill';
-#TODO: spawn an ant and retrieve it's ID
+#spawn an ant and retrieve it's ID
 my $ant = $anthill->ant( 'my ant#1' => [42] );
 is ref($ant), 'Anthill::Ant', 'new ant';
 my $id = $ant->id;
 is $id, 1, 'ant id';
 $ant->start_args( perl_command => {
-			worker => qq^-MAnthill -Ilib -E"\$a=Anthill->new('temp.db');\$a->ant($id)->result(32);say 'hello from ant!'"^,
-			#~ worker => q^-E"say 'hello from ant!'"^,
+			worker => qq^-MAnthill -Ilib -E"\$a=Anthill->new('$dbq');\$a->ant($id)->result(32)->finish;say 'hello from ant!'"^,
 			#~ debug => 1,
 		});
 is $ant->name, 'my ant#1', 'ant name';
@@ -32,6 +34,13 @@ $ant->result({foo => 'bar'});
 is_deeply $ant->result, { foo => 'bar' }, 'ant result';
 
 $ant->wait();
+
+#creating other ants
+$anthill->ant( 'my ant#99'.$_ => [ $_ ] ) for 1..4;
+#querying ants
+my @ants = $anthill->ants( state => [ 'inactive', 'finished' ], args => {like => ($is_sqlite ? '[4%' : '[[]4%') } );
+is scalar @ants, 2, "list ants using filters";
+
 $dbh->disconnect;
-unlink $db;
+unlink $db if $is_sqlite;
 done_testing;
